requirements.txt
langchain_core
python-dotenv
streamlit
langchain_community
faiss-cpu
langchain-groq
langchain-google-genai
google-generativeai
beautifulsoup4